+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.1.1
 
+- Bugfix: reconnect to the SGP30 device on I/O failure.
+
 # 0.1.0
 
 - Initial release.